+++ engine/src/engine/search/mod.rs
@@ -125,8 +125,8 @@ impl<'s> SearchContext<'s> {
 
 #[derive(Debug, Eq, PartialEq)]
 pub enum SearchScore {
-    Centipawns(i16),
-    Mate(i16),
+    Centipawns(i32),
+    Mate(i32),
 }
 
 #[derive(Default)]
