+++ engine/src/engine/search/negamax.rs
@@ -140,7 +140,7 @@ pub fn negamax(
     if !is_root && !is_pv && !in_check {
         // Reverse futility pruning
         if depth <= params::REVERSE_FUTILITY_PRUNE_DEPTH
-            && eval - params::REVERSE_FUTILITY_PRUNE_MARGIN_PER_PLY * i16::from(depth) > beta
+            && eval - params::REVERSE_FUTILITY_PRUNE_MARGIN_PER_PLY * i32::from(depth) > beta
         {
             return Ok(beta);
         }
